Extracted main game loop
