+++ desktop/src/com/nes/tetris/desktop/DesktopLauncher.java
@@ -5,10 +5,10 @@ import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
 import com.nes.tetris.NesTetris;
 
 public class DesktopLauncher {
-	public static void main (String[] arg) {
-		LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
-		config.height = 672;
-		config.width = 768;
-		new LwjglApplication(new NesTetris(), config);
-	}
+    public static void main(String[] arg) {
+        LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
+        config.height = 672;
+        config.width = 768 * 2;
+        new LwjglApplication(new NesTetris(), config);
+    }
 }
